(refactor) Move initalstate of vote status to it's own file
