+++ client/src/data/nominees.js
@@ -3,16 +3,19 @@ const nominees = [
   {
     locationName: "Place1",
     locationImg: "http://placehold.it/200x200",
+    netVotes: 0,
     locationInfo: "http://www.google.com",
   },
   {
     locationName: "Place4",
     locationImg: "http://placehold.it/200x200",
+    netVotes: 0,
     locationInfo: "http://www.google.com",
   },
   {
     locationName: "Place3",
     locationImg: "http://placehold.it/200x200",
+    netVotes: 0,
     locationInfo: "http://www.google.com",
   },
 ];
