+++ client/src/data/voteStatus.js
@@ -1,0 +1,7 @@
+const INITIAL_STATE = {
+  id: '',
+  isVoting: false,
+  winningResult: false
+}
+
+export default INITIAL_STATE;
