+++ client/src/reducers/reducer_userEvents.js
@@ -1,4 +1,4 @@
-import * as types from '../actions/actionCreators';
+import * as types from '../actions/actionTypes';
 // Would be handling all events a user is involved in
   //might need to change the initial state here according to the types of object we get back
     //Also should figure out if we should order these by date
